nits

+++ stream/src/utils/codec.rs
@@ -55,15 +55,15 @@ pub async fn send_frame<S: Sink>(
 /// stream is closed.
 pub async fn recv_frame<T: Stream>(stream: &mut T, max_message_size: u32) -> Result<IoBufs, Error> {
     let (len, skip) = recv_length(stream).await?;
-    if len > max_message_size {
-        return Err(Error::RecvTooLarge(len as usize));
+    if len > max_message_size as usize {
+        return Err(Error::RecvTooLarge(len));
     }
 
     stream
-        .recv(skip as usize + len as usize)
+        .recv(skip + len)
         .await
         .map(|mut bufs| {
-            bufs.advance(skip as usize);
+            bufs.advance(skip);
             bufs
         })
         .map_err(Error::RecvFailed)
@@ -73,7 +73,7 @@ pub async fn recv_frame<T: Stream>(stream: &mut T, max_message_size: u32) -> Res
 /// Returns (payload_len, bytes_to_skip) where bytes_to_skip is:
 /// - varint_len if decoded from peek buffer (bytes not yet consumed)
 /// - 0 if decoded via recv (bytes already consumed)
-async fn recv_length<T: Stream>(stream: &mut T) -> Result<(u32, u32), Error> {
+async fn recv_length<T: Stream>(stream: &mut T) -> Result<(usize, usize), Error> {
     let mut decoder = Decoder::<u32>::new();
 
     // Fast path: decode from peek buffer without blocking
@@ -81,7 +81,7 @@ async fn recv_length<T: Stream>(stream: &mut T) -> Result<(u32, u32), Error> {
         let peeked = stream.peek(MAX_U32_VARINT_SIZE);
         for (i, byte) in peeked.iter().enumerate() {
             match decoder.feed(*byte) {
-                Ok(Some(len)) => return Ok((len, i as u32 + 1)),
+                Ok(Some(len)) => return Ok((len as usize, i + 1)),
                 Ok(None) => continue,
                 Err(_) => return Err(Error::InvalidVarint),
             }
@@ -95,7 +95,7 @@ async fn recv_length<T: Stream>(stream: &mut T) -> Result<(u32, u32), Error> {
 
     loop {
         match decoder.feed(buf.get_u8()) {
-            Ok(Some(len)) => return Ok((len, 0)),
+            Ok(Some(len)) => return Ok((len as usize, 0)),
             Ok(None) => {}
             Err(_) => return Err(Error::InvalidVarint),
         }
